Test: Breaks edge case tests out of test_contract
